backend/requirements/base.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0
celery>=5.4,<6.0
redis>=5.0,<6.0
whitenoise>=6.7,<7.0
django-environ>=0.11
